Add authorization and scope entries to user

+++ app/Http/Controllers/EntriesController.php
@@ -9,7 +9,7 @@ class EntriesController extends Controller
 {
     public function index()
     {
-        $entries = Entry::all()->sortByDesc('created_at');
+        $entries = auth()->user()->entries->sortByDesc('created_at');
 
         return view('entries.index', ['entries' => $entries]);
     }
@@ -26,7 +26,7 @@ class EntriesController extends Controller
             'notes' => 'nullable'
         ]);
 
-        Entry::create($attributes);
+        auth()->user()->entries()->create($attributes);
 
         return redirect('/entries')->with([
             'status' => 'Entry added!'
@@ -35,16 +35,22 @@ class EntriesController extends Controller
 
     public function show(Entry $entry)
     {
+        $this->authorize('manage', $entry);
+
         return view('entries.show', ['entry' => $entry]);
     }
 
     public function edit(Entry $entry)
     {
+        $this->authorize('manage', $entry);
+
         return view('entries.edit', ['entry' => $entry]);
     }
 
     public function update(Request $request, Entry $entry)
     {
+        $this->authorize('manage', $entry);
+
         $attributes = $request->validate([
             'mood' => 'required',
             'notes' => 'nullable'
@@ -59,6 +65,8 @@ class EntriesController extends Controller
 
     public function destroy(Entry $entry)
     {
+        $this->authorize('manage', $entry);
+
         $entry->delete();
 
         return redirect('/entries')->with([
